backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.133.1",
    "fastmcp>=3.0.2",
    "httpx>=0.27.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.13.1",
    "PyJWT[crypto]>=2.9.0",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
]
